create admin approval for deposit money request  for crypto  wallet done

+++ src/main/java/com/oodles/DTO/FiatDepositDTO.java
@@ -2,7 +2,7 @@ package com.oodles.DTO;
 
 public class FiatDepositDTO {
 private String walletType;
-private Long amount;
+private Double amount;
 private Long userId;
 public String getWalletType() {
 	return walletType;
@@ -10,10 +10,11 @@ public String getWalletType() {
 public void setWalletType(String walletType) {
 	this.walletType = walletType;
 }
-public Long getAmount() {
+
+public Double getAmount() {
 	return amount;
 }
-public void setAmount(Long amount) {
+public void setAmount(Double amount) {
 	this.amount = amount;
 }
 public Long getUserId() {
